♻️  ::LectureListDetailViewㄹ를 따로 분리
